--- a/ed25519_test.go
+++ b/ed25519_test.go
@@ -15,7 +15,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/agl/ed25519/edwards25519"
+	"github.com/tendermint/ed25519/edwards25519"
 )
 
 type zeroReader struct{}