This unrolls FeZero and FeOne. This doesn't change the output of the functions, but increases its speed:
```
$ benchcmp old.txt new.txt
benchmark                    old ns/op     new ns/op     delta
BenchmarkKeyGeneration-8     100659        99175         -1.47%
BenchmarkSigning-8           102219        100941        -1.25%
BenchmarkVerification-8      278957        275965        -1.07%
BenchmarkKeyGeneration-8     120965        119236        -1.43%
BenchmarkMap-8               36722         36478         -0.66%
```

The idea was taken from https://github.com/agl/ed25519/pull/11