@cwgoes 2019-01-03
Does Golang's compiler really not do this?