@ValarDragon 2019-01-03
Theres a lot of fighting the compiler you have to do for performant golang code :(, the compiler doesn't do many optimizations for you. (See https://www.youtube.com/watch?v=keydVd-Zn80 for the struggles of in-lining)